# specs/001-logistics-bill-app/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Logistics Delivery Bill Management API
  version: 0.1.0
  summary: Backend HTTP contract for the Vietnamese logistics delivery-bill app.
  description: |
    Phase 1 contract for the v1 scope: authentication, customer profiles,
    delivery bills (create, search, detail, status update, print, reprint, PDF export),
    and the service-tier catalogue.

    All free-text request and response fields contain Vietnamese with full diacritics.
    Money values are JSON numbers in VND (no currency symbol). Weights are in kg
    with up to 3 decimal places. Dimensions are in cm with up to 2 decimal places.
    Numeric fields are transmitted with a dot decimal separator over the wire,
    even though the UI displays them with a Vietnamese comma separator.

servers:
  - url: /api/v1
    description: Default mount point of the FastAPI router.

tags:
  - name: auth
    description: Đăng nhập / Authentication
  - name: bills
    description: Phiếu Gửi / Delivery bills
  - name: customers
    description: Khách hàng / Customer profiles
  - name: service-tiers
    description: Dịch vụ / Service tier catalogue

security:
  - bearerAuth: []

paths:

  /auth/login:
    post:
      tags: [auth]
      summary: Đăng nhập
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/Error'

  /auth/refresh:
    post:
      tags: [auth]
      summary: Làm mới token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/Error'

  /auth/me:
    get:
      tags: [auth]
      summary: Thông tin người dùng hiện tại
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Error'

  /service-tiers:
    get:
      tags: [service-tiers]
      summary: Danh sách dịch vụ
      parameters:
        - in: query
          name: scope
          required: false
          schema:
            type: string
            enum: [domestic, international]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceTier'

  /customers:
    get:
      tags: [customers]
      summary: Tra cứu khách hàng
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: q
          required: false
          schema:
            type: string
          description: Tìm theo tên / địa chỉ / SĐT / mã KH (không phân biệt dấu).
        - in: query
          name: is_active
          required: false
          schema: { type: boolean }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerPage'
    post:
      tags: [customers]
      summary: Tạo khách hàng mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'

  /customers/{customer_id}:
    parameters:
      - $ref: '#/components/parameters/CustomerId'
    get:
      tags: [customers]
      summary: Chi tiết khách hàng
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Customer' }
        '404':
          $ref: '#/components/responses/Error'
    patch:
      tags: [customers]
      summary: Cập nhật khách hàng
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Customer' }
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    delete:
      tags: [customers]
      summary: Vô hiệu hoá khách hàng (soft delete)
      responses:
        '204':
          description: No Content
        '404':
          $ref: '#/components/responses/Error'

  /bills:
    get:
      tags: [bills]
      summary: Danh sách phiếu gửi
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: q
          required: false
          schema: { type: string }
          description: Tìm theo mã vận đơn / tên / SĐT (không phân biệt dấu).
        - in: query
          name: status
          required: false
          schema:
            $ref: '#/components/schemas/BillStatus'
        - in: query
          name: customer_code
          required: false
          schema: { type: string }
        - in: query
          name: created_from
          required: false
          schema: { type: string, format: date-time }
        - in: query
          name: created_to
          required: false
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillPage'
    post:
      tags: [bills]
      summary: Tạo phiếu gửi mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BillCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Bill' }
        '400':
          $ref: '#/components/responses/Error'

  /bills/{bill_id}:
    parameters:
      - $ref: '#/components/parameters/BillId'
    get:
      tags: [bills]
      summary: Chi tiết phiếu gửi
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Bill' }
        '404':
          $ref: '#/components/responses/Error'

  /bills/by-tracking/{tracking_number}:
    parameters:
      - in: path
        name: tracking_number
        required: true
        schema: { type: string }
    get:
      tags: [bills]
      summary: Tra cứu theo mã vận đơn
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Bill' }
        '404':
          $ref: '#/components/responses/Error'

  /bills/{bill_id}/status:
    parameters:
      - $ref: '#/components/parameters/BillId'
    post:
      tags: [bills]
      summary: Cập nhật trạng thái phiếu gửi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BillStatusUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Bill' }
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'

  /bills/{bill_id}/print:
    parameters:
      - $ref: '#/components/parameters/BillId'
    get:
      tags: [bills]
      summary: Tải phiếu gửi PDF (in/in lại)
      description: |
        Renders the canonical bill HTML template via WeasyPrint and streams the
        PDF. Each call increments `print_count` and writes a `bill.printed`
        audit event attributed to the caller.
      parameters:
        - in: query
          name: as
          required: false
          schema: { type: string, enum: [pdf, html], default: pdf }
      responses:
        '200':
          description: OK
          content:
            application/pdf: {}
            text/html: {}
        '404':
          $ref: '#/components/responses/Error'

  /bills/{bill_id}/events:
    parameters:
      - $ref: '#/components/parameters/BillId'
    get:
      tags: [bills]
      summary: Lịch sử trạng thái và kiểm toán
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status_events:
                    type: array
                    items: { $ref: '#/components/schemas/BillStatusEvent' }
                  audit_events:
                    type: array
                    items: { $ref: '#/components/schemas/AuditEvent' }
                required: [status_events, audit_events]
        '404':
          $ref: '#/components/responses/Error'

components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    Page:
      in: query
      name: page
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      in: query
      name: page_size
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    BillId:
      in: path
      name: bill_id
      required: true
      schema: { type: integer, format: int64 }
    CustomerId:
      in: path
      name: customer_id
      required: true
      schema: { type: integer, format: int64 }

  responses:
    Error:
      description: Lỗi
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:

    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          type: string
          example: BILL_NOT_FOUND
        message:
          type: string
          description: Thông báo lỗi bằng tiếng Việt.
          example: Không tìm thấy phiếu gửi.
        details:
          type: object
          additionalProperties: true

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }

    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }

    TokenPair:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, enum: [bearer] }
        expires_in:
          type: integer
          description: Seconds until access_token expires.

    User:
      type: object
      required: [id, username, full_name, role, is_active]
      properties:
        id: { type: integer, format: int64 }
        username: { type: string }
        full_name: { type: string }
        role:
          type: string
          enum: [staff, supervisor, admin]
        is_active: { type: boolean }

    ServiceTier:
      type: object
      required: [code, display_name, scope, is_active]
      properties:
        code:
          type: string
          enum:
            [CPN, PHT, DUONG_BO, T48H, NGUYEN_CHUYEN, KHAC,
             INTL_EXPRESS, INTL_ECONOMY, INTL_OTHER]
        display_name: { type: string }
        scope:
          type: string
          enum: [domestic, international]
        is_active: { type: boolean }

    CustomerBase:
      type: object
      required: [display_name, address, district, province, phone]
      properties:
        display_name: { type: string }
        address: { type: string }
        district: { type: string }
        province: { type: string }
        phone:
          type: string
          pattern: '^\+?[0-9 \-]{8,20}$'

    CustomerCreate:
      allOf:
        - $ref: '#/components/schemas/CustomerBase'
        - type: object
          required: [customer_code]
          properties:
            customer_code:
              type: string
              description: E.g. NL500010

    CustomerUpdate:
      allOf:
        - $ref: '#/components/schemas/CustomerBase'
        - type: object
          properties:
            is_active: { type: boolean }

    Customer:
      allOf:
        - $ref: '#/components/schemas/CustomerCreate'
        - type: object
          required: [id, is_active, created_at, updated_at]
          properties:
            id: { type: integer, format: int64 }
            is_active: { type: boolean }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }

    CustomerPage:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Customer' }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }

    Party:
      type: object
      required: [name, address, district, province, phone]
      properties:
        name: { type: string }
        address: { type: string }
        district: { type: string }
        province: { type: string }
        phone:
          type: string
          pattern: '^\+?[0-9 \-]{8,20}$'

    BillContentLine:
      type: object
      required: [description, quantity, weight_kg]
      properties:
        line_no: { type: integer, minimum: 1 }
        description: { type: string }
        quantity: { type: integer, minimum: 1 }
        weight_kg:
          type: number
          minimum: 0
          format: double
        length_cm:
          type: number
          minimum: 0
          format: double
        width_cm:
          type: number
          minimum: 0
          format: double
        height_cm:
          type: number
          minimum: 0
          format: double

    FeeBreakdown:
      type: object
      required: [fee_main, fee_fuel_surcharge, fee_other_surcharge, fee_vat, fee_total]
      properties:
        fee_main: { type: number, format: double, minimum: 0 }
        fee_fuel_surcharge: { type: number, format: double, minimum: 0 }
        fee_other_surcharge: { type: number, format: double, minimum: 0 }
        fee_vat: { type: number, format: double, minimum: 0 }
        fee_total:
          type: number
          format: double
          minimum: 0
          description: |
            Must equal fee_main + fee_fuel_surcharge + fee_other_surcharge + fee_vat.
            Server rejects mismatching totals with HTTP 400.

    BillCreate:
      type: object
      required:
        [sender, receiver, contents, cargo_type, service_tier_code,
         fee, payer]
      properties:
        customer_id: { type: integer, format: int64, nullable: true }
        customer_code:
          type: string
          nullable: true
          description: Snapshot value for "Mã KH"; auto-filled when customer_id set.
        sender: { $ref: '#/components/schemas/Party' }
        receiver: { $ref: '#/components/schemas/Party' }
        contents:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/BillContentLine' }
        cargo_type:
          type: string
          enum: [document, goods]
        service_tier_code:
          type: string
          description: |
            Must reference an active service_tiers.code. Domestic and
            international tiers are mutually exclusive (FR-005).
        fee: { $ref: '#/components/schemas/FeeBreakdown' }
        payer:
          type: string
          enum: [sender, receiver]

    BillStatus:
      type: string
      enum: [da_tao, da_lay_hang, dang_van_chuyen, da_giao, hoan_tra, huy]

    Bill:
      type: object
      required:
        [id, tracking_number, sender, receiver, contents, cargo_type,
         service_tier_code, fee, payer, status, created_at, updated_at,
         created_by, updated_by, print_count]
      properties:
        id: { type: integer, format: int64 }
        tracking_number:
          type: string
          example: NL0011310
        customer_code:
          type: string
          nullable: true
          example: NL500010
        customer_id:
          type: integer
          format: int64
          nullable: true
        sender: { $ref: '#/components/schemas/Party' }
        receiver: { $ref: '#/components/schemas/Party' }
        contents:
          type: array
          items: { $ref: '#/components/schemas/BillContentLine' }
        cargo_type:
          type: string
          enum: [document, goods]
        service_tier_code: { type: string }
        fee: { $ref: '#/components/schemas/FeeBreakdown' }
        payer:
          type: string
          enum: [sender, receiver]
        status: { $ref: '#/components/schemas/BillStatus' }
        delivered_at:
          type: string
          format: date-time
          nullable: true
        delivered_to_name:
          type: string
          nullable: true
        cancellation_reason:
          type: string
          nullable: true
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        created_by: { type: integer, format: int64 }
        updated_by: { type: integer, format: int64 }
        print_count: { type: integer }
        last_printed_at:
          type: string
          format: date-time
          nullable: true
        last_printed_by:
          type: integer
          format: int64
          nullable: true

    BillPage:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Bill' }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }

    BillStatusUpdate:
      type: object
      required: [to_status]
      properties:
        to_status: { $ref: '#/components/schemas/BillStatus' }
        delivered_to_name:
          type: string
          description: Bắt buộc khi to_status = da_giao.
        cancellation_reason:
          type: string
          description: Bắt buộc khi to_status = huy.
        note: { type: string }

    BillStatusEvent:
      type: object
      required: [id, bill_id, to_status, actor_id, created_at]
      properties:
        id: { type: integer, format: int64 }
        bill_id: { type: integer, format: int64 }
        from_status:
          $ref: '#/components/schemas/BillStatus'
        to_status:
          $ref: '#/components/schemas/BillStatus'
        note:
          type: string
          nullable: true
        actor_id: { type: integer, format: int64 }
        created_at: { type: string, format: date-time }

    AuditEvent:
      type: object
      required: [id, action, created_at]
      properties:
        id: { type: integer, format: int64 }
        actor_id:
          type: integer
          format: int64
          nullable: true
        action: { type: string }
        entity_type:
          type: string
          nullable: true
        entity_id:
          type: integer
          format: int64
          nullable: true
        details:
          type: object
          additionalProperties: true
          nullable: true
        created_at: { type: string, format: date-time }